Don't encode spaces as pluses
I've come to accept that this is bad practice. Spaces as pluses
should *only* appear in query strings, where they are a historical
accident coming from a hack to make search queries easier to
compose. The fact that they look *slightly* nicer than %20 is not
a valid reason to re-introduce their complexity.